reduced Sandwiches Factory Method exercise to fewer files
