--- Design-Patterns/Factory-Method/Ingredients/Bread.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Factory_Method.Ingredients
-{
-    internal class Bread : Ingredient
-    {
-
-    }
-}
