+++ Design-Patterns/Factory-Method/Program.cs
@@ -1,5 +1,4 @@
-﻿using Factory_Method.Sandwiches;
-
+﻿
 namespace Factory_Method
 {
     internal class Program
