--- Design-Patterns/Factory-Method/Sandwiches/TurkeySandwich.cs
@@ -1,22 +1,0 @@
-﻿using Factory_Method.Ingredients;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Factory_Method.Sandwiches
-{
-    internal class TurkeySandwich : Sandwich
-    {
-        public override void CreateIngredients()
-        {
-            Ingredients.Add(new Bread());
-            Ingredients.Add(new Mayonnaise());
-            Ingredients.Add(new Turkey());
-            Ingredients.Add(new Turkey());
-            Ingredients.Add(new Lettuce());
-            Ingredients.Add(new Bread());
-        }
-    }
-}
